feat(Update): project manager list show all, latest activities, tab navigation on scope & credentials, module new status, filter progress, dynamic version progress

+++ resources/views/project/project_manager/pages/project/index.blade.php
@@ -5,6 +5,8 @@
 @section('style')
     <link rel="stylesheet" href="{{ asset('templates/stisla/node_modules/datatables.net-bs4/css/dataTables.bootstrap4.min.css') }}">
     <link rel="stylesheet" href="{{ asset('templates/stisla/node_modules/datatables.net-select-bs4/css/select.bootstrap4.min.css') }}">
+    <link rel="stylesheet" href="{{ asset('templates/stisla/node_modules/select2/dist/css/select2.min.css') }}"/>
+    <link rel="stylesheet" href="{{ asset('templates/stisla/node_modules/bootstrap-daterangepicker/daterangepicker.css') }}"/>
 @endsection
 
 @section('content')
@@ -24,34 +26,69 @@
                     <form role="form" method="GET">
                         <div class="form-row">
                             <div class="form-group col-md-3 col-lg-3">
-                                <label>Start Date</label>
-                                <input type="date" class="form-control" name="start_date" value="{{ $request->start_date }}">
-                            </div>
-                            <div class="form-group col-md-3 col-lg-3">
-                                <label>End Date</label>
-                                <input type="date" class="form-control" name="end_date" value="{{ $request->end_date }}">
+                                <label>Start - End Date</label>
+                                <div class="input-group">
+                                    <div class="input-group-prepend">
+                                        <div class="input-group-text">
+                                            <i class="fas fa-calendar"></i>
+                                        </div>
+                                    </div>
+                                    <input
+                                      type="text"
+                                      name="start_end_date"
+                                      class="form-control daterange-cus"
+                                      value="{{ $request->start_end_date }}"
+                                    />
+                                </div>
                             </div>
                             <div class="form-group col-md-3 col-lg-3">
                                 <label>User Assignee</label>
-                                <select class="form-control form-control-sm" name="user">
-                                    <option value="no_filter">All</option>
-                                    @foreach($roles as $role)
-                                        @foreach($role->user as $user)
-                                            <option value="{{ $user->id }}" @if($user->id == $request->user) selected @endif>{{ $user->name }}</option>
-                                        @endforeach
+                                <select class="form-control select2" name="user[]" multiple="">
+                                    @foreach($employees as $employee)
+                                        <option value="{{ $employee->id }}"
+                                            @if(!empty($request->user))
+                                                @foreach($request->user as $requested)
+                                                    @if($employee->id == $requested)
+                                                        selected
+                                                    @endif
+                                                @endforeach
+                                            @endif
+                                        >
+                                            {{ $employee->name }}
+                                        </option>
                                     @endforeach
                                 </select>
                             </div>
-                            <div class="form-group col-md-3 col-lg-2">
+                            <div class="form-group col-md-3 col-lg-3">
+                                <label>Project Manager</label>
+                                <select class="form-control select2" name="project_manager">
+                                    <option value="no_filter">All</option>
+                                    @foreach($projectManagers as $projectManager)
+                                        <option value="{{ $projectManager->id }}" @if($projectManager->id == $request->project_manager) selected @endif>
+                                            {{ $projectManager->name }}
+                                        </option>
+                                    @endforeach
+                                </select>
+                            </div>
+                            <div class="form-group col-md-2 col-lg-2">
                                 <label>Role Assignee</label>
-                                <select class="form-control form-control-sm" name="role">
-                                    <option value="no_filter">All</option>
+                                <select class="form-control select2" name="role[]" multiple="">
                                     @foreach($roles as $role)
-                                        <option value="{{ $role->id }}" @if($role->id == $request->role) selected @endif>{{ $role->name }}</option>
+                                        <option value="{{ $role->id }}" 
+                                            @if(!empty($request->role))
+                                                @foreach($request->role as $requested)
+                                                    @if($role->id == $requested)
+                                                        selected
+                                                    @endif
+                                                @endforeach
+                                            @endif
+                                        >
+                                            {{ $role->name }}
+                                        </option>
                                     @endforeach
                                 </select>
                             </div>
-                            <div class="form-group col-md-4 col-lg-1">
+                            <div class="form-group col-md-1 col-lg-1">
                                 <button type="submit" class="btn btn-lg btn-primary position-absolute" style="bottom: 0"><i class="fa fa-search"></i></button>
                             </div>
                         </div>
@@ -69,39 +106,44 @@
                             </thead>
                             <tbody>
                                 @foreach($projects as $project)
-                                    <tr>
-                                        <td>
-                                            {{ $project->name }}
-                                        </td>
-                                        <td>
-                                            {{ $project->start_date->format('d-m-Y') }}
-                                        </td>
-                                        <td>
-                                            {{ $project->end_date->format('d-m-Y') }}
-                                        </td>
-                                        <td>
-                                            @if($project->projectVersions->count() > 1)
-                                                <div class="badge badge-info">Maintenance</div>
-                                            @elseif(!empty($project->launch_date))
-                                                <div class="badge badge-success">Launch</div>
-                                            @elseif($project->projectVersions->last()->projectDetails()->whereDoneOrOnProgress()->count() > 0)
-                                                <div class="badge badge-warning">Development</div>
-                                            @else
-                                                <div class="badge badge-danger">Listed</div>
-                                            @endif
-                                        </td>
-                                        <td>
-                                            <a href="{{ route('project_manager.projects.detail', $project) }}" class="btn btn-light mr-1" data-toggle="tooltip" title="View"><i class="fas fa-eye"></i></a>
-                                            <a href="{{ route('project_manager.projects.edit', $project->id) }}" class="btn btn-primary btn-action mr-1" data-toggle="tooltip" title="Edit"><i class="fas fa-pencil-alt"></i></a>
-                                            <a href="#" onclick="deleteConfirm('del{{ $project->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
-                                                <i class="fa fa-trash"></i>
-                                            </a>
-                                            <form id="del{{ $project->id }}" action="{{ route('project_manager.projects.destroy', $project) }}" method="POST">
-                                                @method('delete')
-                                                @csrf
-                                            </form>
-                                        </td>
-                                    </tr>
+                                    {{-- @if(!empty($request->user) && !array_diff($request->user, $project->userAssignments()->pluck('user_id')->toArray())) --}}
+                                        <tr>
+                                            <td>
+                                                {{ $project->name }}
+                                            </td>
+                                            <td>
+                                                {{ $project->start_date->format('d-m-Y') }}
+                                            </td>
+                                            <td>
+                                                {{ $project->end_date->format('d-m-Y') }}
+                                            </td>
+                                            <td>
+                                                @if($project->projectVersions->count() > 1)
+                                                    <div class="badge badge-info">Maintenance</div>
+                                                @elseif(!empty($project->launch_date))
+                                                    <div class="badge badge-success">Launch</div>
+                                                @elseif($project->projectVersions->last()->projectDetails()->whereDoneOrOnProgress()->count() > 0)
+                                                    <div class="badge badge-warning">Development</div>
+                                                @else
+                                                    <div class="badge badge-danger">Listed</div>
+                                                @endif
+                                            </td>
+                                            <td>
+                                                <a href="{{ route('project_manager.projects.detail', $project) }}" class="btn btn-light mr-1" data-toggle="tooltip" title="View"><i class="fas fa-eye"></i></a>
+
+                                                @if($project->user_id == Auth::user()->id)
+                                                    <a href="{{ route('project_manager.projects.edit', $project->id) }}" class="btn btn-primary btn-action mr-1" data-toggle="tooltip" title="Edit"><i class="fas fa-pencil-alt"></i></a>
+                                                    <a href="#" onclick="deleteConfirm('del{{ $project->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
+                                                        <i class="fa fa-trash"></i>
+                                                    </a>
+                                                    <form id="del{{ $project->id }}" action="{{ route('project_manager.projects.destroy', $project) }}" method="POST">
+                                                        @method('delete')
+                                                        @csrf
+                                                    </form>
+                                                @endif
+                                            </td>
+                                        </tr>
+                                    {{-- @endif --}}
                                 @endforeach
                             </tbody>
                         </table>
@@ -117,6 +159,15 @@
     <script src="{{ asset('templates/stisla/node_modules/datatables.net-bs4/js/dataTables.bootstrap4.min.js') }}"></script>
     <script src="{{ asset('templates/stisla/node_modules/datatables.net-select-bs4/js/select.bootstrap4.min.js') }}"></script>
     <script src="{{ asset('templates/stisla/node_modules/sweetalert/dist/sweetalert.min.js') }}"></script>
+    <script src="{{ asset('templates/stisla/node_modules/select2/dist/js/select2.full.min.js') }}"></script>
+    <script src="{{ asset('templates/stisla/node_modules/bootstrap-daterangepicker/daterangepicker.js') }}"></script>
+    <script>
+        $('.daterange-cus').daterangepicker({
+            locale: {format: 'YYYY-MM-DD'},
+            drops: 'down',
+            opens: 'right'
+        });
+    </script>
     @if (Session::has('success'))
         <script>
             swal("Success!", "{{ Session::get('success') }}", "success").then(function(){
